requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
betterproto==2.0.0b7
black==25.1.0
blockbuster==1.5.24
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
cryptography==44.0.3
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastmcp==2.2.10
filelock==3.18.0
fireworks-ai==0.16.1
flatbuffers==25.2.10
forbiddenfruit==0.1.4
frozenlist==1.6.0
fsspec==2025.3.2
googlemaps==4.10.0
greenlet==3.2.1
grpcio==1.67.1
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hf-xet==1.1.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.31.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-anthropic==0.3.12
langchain-community==0.3.23
langchain-core==0.3.61
langchain-fireworks==0.3.0
langchain-mcp-adapters==0.0.10
langchain-openai==0.3.18
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-api==0.2.21
langgraph-checkpoint==2.0.25
langgraph-cli==0.2.10
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.11
langgraph-sdk==0.1.66
langgraph-supervisor==0.0.22
langsmith==0.3.42
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.7.1
mdurl==0.1.2
milvus-lite==2.4.12
mpmath==1.3.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.21.1
openai==1.77.0
openapi-pydantic==0.5.1
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.30.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyExifTool==0.5.6
Pygments==2.19.1
PyJWT==2.10.1
pymilvus==2.5.8
pymilvus.model==0.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.3.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
truststore==0.10.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
